requirements.txt
semantic-kernel
colorama
azure-identity
selenium